+++ v03_pipeline/lib/annotations/snv.py
@@ -27,3 +27,36 @@ def DP(mt: hl.MatrixTable, **_: Any) -> hl.Expression:  # noqa: N802
         hl.int(hl.min(hl.sum(mt.AD), 32000)),
         hl.missing(hl.tint),
     )
+
+
+def gnomad_non_coding_constraint(
+    ht: hl.Table,
+    interval_ht: hl.Table,
+    **_: Any,
+) -> hl.Expression:
+    return hl.Struct(
+        z_score=(
+            interval_ht.index(ht.locus, all_matches=True)
+            .filter(
+                lambda x: hl.is_defined(x.gnomad_non_coding_constraint['z_score']),
+            )
+            .gnomad_non_coding_constraint.z_score.first()
+        ),
+    )
+
+
+def screen(
+    ht: hl.Table,
+    interval_ht: hl.Table,
+    **_: Any,
+) -> hl.Expression:
+    return hl.Struct(
+        region_type_ids=(
+            interval_ht.index(
+                ht.locus,
+                all_matches=True,
+            ).flatmap(
+                lambda x: x.screen['region_type_ids'],
+            )
+        ),
+    )
